+++ telegram_click/util.py
@@ -19,11 +19,8 @@
 #  SOFTWARE.
 
 import logging
-from collections import OrderedDict
 
 from telegram import Bot
-
-from telegram_click.const import ARG_NAMING_PREFIXES
 
 LOGGER = logging.getLogger(__name__)
 
@@ -74,173 +71,6 @@ def escape_for_markdown(text: str or None) -> str:
     text = str(text)
     escaped = text.replace("*", "\\*").replace("_", "\\_")
     return escaped
-
-
-def starts_with_naming_prefix(arg: str) -> bool:
-    for prefix in ARG_NAMING_PREFIXES:
-        if arg.startswith(prefix):
-            return True
-
-    return False
-
-
-def remove_naming_prefix(arg: str) -> str:
-    for prefix in ARG_NAMING_PREFIXES:
-        if arg.startswith(prefix):
-            return arg[len(prefix):]
-
-    return arg
-
-
-def parse_command_args(arguments: str or None, expected_args: []) -> dict:
-    """
-    Parses the given argument text
-    :param arguments: the argument text
-    :param expected_args: a list of expected arguments
-    :return: dictionary { argument-name -> value }
-    """
-    if arguments is None:
-        arguments = ""
-
-    import shlex
-    str_args = shlex.split(arguments)
-
-    # map argument.name -> argument
-    arg_name_map = OrderedDict()
-    for expected_arg in expected_args:
-        for name in expected_arg.names:
-            arg_name_map[name] = expected_arg
-
-    parsed_args = {}
-
-    named_arg_idx = []
-    for idx, val in enumerate(str_args):
-        if starts_with_naming_prefix(val):
-            named_arg_idx.append(idx)
-
-    # process named arguments first
-    used_idx = list(named_arg_idx)
-    for idx in named_arg_idx:
-        val = str_args[idx]
-        arg_name = remove_naming_prefix(val)
-        if arg_name not in arg_name_map:
-            raise ValueError("Unknown argument '{}'".format(arg_name))
-        arg = arg_name_map[arg_name]
-
-        if arg.flag:
-            # since it is present, we assume a true value
-            value = arg.parse_arg_value("True")
-        else:
-            value = str_args[idx + 1]
-            used_idx.append(idx + 1)
-            if starts_with_naming_prefix(value):
-                raise ValueError(
-                    "Expected argument value for '--{}' but found named argument '{}'".format(arg_name, value))
-
-        parsed_args[arg.name] = arg.parse_arg_value(value)
-        for name in arg.names:
-            arg_name_map.pop(name)
-
-    remaining_idx = list(set(list(range(len(str_args)))) - set(used_idx))
-    for idx in sorted(remaining_idx):
-        if len(arg_name_map) <= 0:
-            # ignore excess arguments
-            break
-
-        val = str_args[idx]
-        arg = list(arg_name_map.values())[0]
-        parsed_args[arg.name] = arg.parse_arg_value(val)
-        for name in arg.names:
-            arg_name_map.pop(name)
-
-    # and then handle missing args
-    while len(arg_name_map) > 0:
-        name, arg = arg_name_map.popitem()
-        parsed_args[arg.name] = arg.parse_arg_value(None)
-        for name in list(filter(lambda x: x != name, arg.names)):
-            arg_name_map.pop(name)
-
-    return parsed_args
-
-
-def split_command_from_args(text: str or None) -> (str or None, str or None):
-    """
-    Splits the command (including any target) from its arguments
-    :param text: the full message
-    :return: (command, args)
-    """
-    if text is None or len(text) <= 0:
-        return None, None
-
-    if " " in text:
-        return text.split(" ", 1)
-    else:
-        return text, None
-
-
-def split_command_from_target(bot_username: str, command: str or None) -> (str, str):
-    """
-    Determines the command target bot username
-    :param bot_username: the username of this bot
-    :param command: the command to check
-    :return: (the command, the username of the targeted bot)
-    """
-    target = bot_username
-
-    if command is None or len(command) <= 0:
-        return command, target
-
-    if '@' in command:
-        command, target = command.split('@', 1)
-
-    return command, target
-
-
-def parse_telegram_command(bot_username: str, text: str, expected_args: []) -> (str, str, [str]):
-    """
-    Parses the given message to a command and its arguments
-    :param bot_username: the username of the current bot
-    :param text: the text to parse
-    :param expected_args: expected arguments
-    :return: the target bot username, command, and its argument list
-    """
-    command, args = split_command_from_args(text)
-    command, target = split_command_from_target(bot_username, command)
-    parsed_args = parse_command_args(args, expected_args)
-    return command[1:], parsed_args
-
-
-def generate_help_message(names: [str], description: str, args: []) -> str:
-    """
-    Generates a command usage description
-    :param names: names of the command
-    :param description: command description
-    :param args: command argument list
-    :return: help message
-    """
-    command_names = list(map(lambda x: "/{}".format(escape_for_markdown(x)), names))
-    command_names_line = command_names[0]
-    if len(command_names) > 1:
-        command_names_line += " ({})".format(", ".join(command_names[1:]))
-
-    argument_lines = list(map(lambda x: x.generate_argument_message(), args))
-    arguments = "\n".join(argument_lines)
-
-    argument_examples = " ".join(list(map(lambda x: x.example, args)))
-
-    lines = [
-        command_names_line,
-        description
-    ]
-    if len(arguments) > 0:
-        lines.extend([
-            "Arguments: ",
-            arguments,
-            "Example:",
-            "  `/{} {}`".format(names[0], argument_examples)
-        ])
-
-    return "\n".join(lines)
 
 
 def send_message(bot: Bot, chat_id: str, message: str, parse_mode: str = None, reply_to: int = None):
